+++ python/graphframes/tests.py
@@ -84,6 +84,10 @@ class GraphFrameLibTest(GraphFrameTestCase):
         self.sqlContext = self.sql
         self.japi = _java_api(self.sqlContext._sc)
 
+    def _hasCols(self, graph, vcols = [], ecols = []):
+        graph.vertices.select(*vcols).collect()
+        graph.edges.select(*ecols).collect()
+
     def _graph(self, name, *args):
         """
         Convenience to call one of the example graphs, passing the arguments and wrapping the result back
@@ -101,7 +105,8 @@ class GraphFrameLibTest(GraphFrameTestCase):
         (0L, "a", "b")], ["id", "vattr", "gender"])
         e = self.sqlContext.createDataFrame([(0L, 0L, 1L)], ["src", "dst", "test"]).filter("src > 10")
         g = GraphFrame(v, e)
-        comps = g.connected_components()
+        comps = g.connectedComponents()
+        self._hasCols(comps, vcols=['component', 'vattr', 'gender'])
         self.assertEqual(comps.vertices.count(), 1)
         self.assertEqual(comps.edges.count(), 0)
 
@@ -109,14 +114,15 @@ class GraphFrameLibTest(GraphFrameTestCase):
         v = self.sqlContext.createDataFrame([(0L, "a0", "b0"), (1L, "a1", "b1")], ["id", "A", "B"])
         e = self.sqlContext.createDataFrame([(0L, 1L, "a01", "b01")], ["src", "dst", "A", "B"])
         g = GraphFrame(v, e)
-        comps = g.connected_components()
+        comps = g.connectedComponents()
+        self._hasCols(comps, vcols=['component', 'A', 'B'])
         self.assertEqual(comps.vertices.count(), 2)
         self.assertEqual(comps.edges.count(), 1)
 
     def test_label_progagation(self):
         n = 5
         g = self._graph("twoBlobs", n)
-        labels = g.label_propagation(max_steps=4 * n)
+        labels = g.labelPropagation(max_steps=4 * n)
         labels1 = labels.vertices.filter("id < 5").select("label").collect()
         all1 = set([x.label for x in labels1])
         self.assertEqual(all1, set([0]))
@@ -129,7 +135,8 @@ class GraphFrameLibTest(GraphFrameTestCase):
         g = self._graph("star", n)
         resetProb = 0.15
         errorTol = 1.0e-5
-        pr = g.page_rank(resetProb, tolerance=errorTol)
+        pr = g.pageRank(resetProb, tolerance=errorTol)
+        self._hasCols(pr, vcols=['weight'], ecols=['weight'])
 
     def test_shortest_paths(self):
         edges = [(1, 2), (1, 5), (2, 3), (2, 5), (3, 4), (4, 5), (4, 6)]
@@ -138,19 +145,21 @@ class GraphFrameLibTest(GraphFrameTestCase):
         vertices = self.sqlContext.createDataFrame([(i,) for i in range(1, 7)], ["id"])
         g = GraphFrame(vertices, edges)
         landmarks = [1, 4]
-        g2 = g.shortest_paths(landmarks)
+        g2 = g.shortestPaths(landmarks)
         rows = g2.vertices.select("id", "distance").collect()
 
     def test_svd_plus_plus(self):
         g = self._graph("ALSSyntheticData")
-        (g2, cost) = g.svd_plus_plus({})
+        (g2, cost) = g.svdPlusPlus({})
+        self._hasCols(g2, vcols=['column1', 'column2', 'column3'], ecols=['ecolumn1'])
+
 
     def test_strongly_connected_components(self):
         # Simple island test
         vertices = self.sqlContext.createDataFrame([(i,) for i in range(1, 6)], ["id"])
         edges = self.sqlContext.createDataFrame([(7, 8)], ["src", "dst"])
         g = GraphFrame(vertices, edges)
-        c = g.strongly_connected_components(5)
+        c = g.stronglyConnectedComponents(5)
         for row in c.vertices.collect():
             self.assertEqual(row.id, row.component)
 
@@ -158,7 +167,7 @@ class GraphFrameLibTest(GraphFrameTestCase):
         edges = self.sqlContext.createDataFrame([(0, 1), (1, 2), (2, 0)], ["src", "dst"])
         vertices = self.sqlContext.createDataFrame([(0,), (1,), (2,)], ["id"])
         g = GraphFrame(vertices, edges)
-        c = g.triangle_count()
+        c = g.triangleCount()
         for row in c.vertices.select("id", "count").collect():
             self.assertEqual(row.count, 1)
